backend/requirements.txt
alembic==1.13.2
bidict==0.23.1
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.2
flask-sqlalchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
importlib-metadata==8.3.0
importlib-resources==6.4.3
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
SQLAlchemy==2.0.32
typing-extensions==4.12.2
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.20.0
